#74 feat: adapatation result added

+++ adaptation/api/serializers.py
@@ -91,14 +91,16 @@ class AdaptationClosedUpdateSerializer(serializers.ModelSerializer, ErrorNameMix
 
     class Meta:
         model = Adaptation
-        fields = ["is_closed"]
+        fields = ['is_closed', 'result_note']
 
     def update(self, instance, validated_data):
         with transaction.atomic():
-
+            is_closed = validated_data.get('is_closed', None)
+            if is_closed is not None:
+                if not instance.is_closed:
+                    instance.confirmations.all().delete()
             data = super().update(instance, validated_data) 
-            if not instance.is_closed:
-                instance.confirmations.all().delete()
+            print(data)
         return data
